chore: add dynamic footer
