+++ src/assets/translations/en/main.json
@@ -130,6 +130,7 @@
     "best": "Best",
     "gasFee": "Gas Fee",
     "approvalFee": "Approval Fee",
+    "allowanceResetFee": "Allowance Reset Fee",
     "consolidationFee": "Consolidation Fee",
     "fees": "Fees",
     "alternative": "Alternative",
